% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agents.R
\name{agents}
\alias{agents}
\title{Agents}
\usage{
agents(id = NULL, api_url = "https://eventdata.datacite.org", ...)
}
\arguments{
\item{id}{(character) agent ID. optional}

\item{api_url}{(character) API endpoint, defaults to http://alm.plos.org/api/v5/articles}

\item{...}{Curl options passed on to \code{\link[httr]{GET}}}
}
\description{
Agents
}
\examples{
\dontrun{
res <- agents()
res$meta
head(res$agents)

agents(id = 'plos_fulltext')
}
}

