% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/alerts.R
\name{alerts}
\alias{alerts}
\title{API alerts}
\usage{
alerts(key, api_url = "http://alm.plos.org", ...)
}
\arguments{
\item{key}{(character) An API key. Required.}

\item{api_url}{(character) API endpoint, defaults to http://alm.plos.org/api/v5/articles}

\item{...}{Curl options (debugging tools mostly) passed on to \code{\link[httr]{GET}}}
}
\description{
API alerts
}
\examples{
\dontrun{
alerts(getOption("almv4_pwd"))
}
}

