% Generated by roxygen2 (4.0.1): do not edit by hand
\name{alm_pubmedid}
\alias{alm_pubmedcentid}
\alias{alm_pubmedid}
\alias{alm_totals}
\alias{almdateupdated}
\alias{almupdated}
\title{Get PubMed article ID by inputting the doi for the article.}
\usage{
alm_pubmedid(...)

alm_pubmedcentid(...)

almupdated(...)

almdateupdated(...)

alm_totals(doi, key = NULL)
}
\description{
Get PubMed article ID by inputting the doi for the article.

Get PubMed Central article ID by inputting the doi for the article.

Date when alt-metrics for the article (by DOI) data was last updated.

Get the date when article was last updated.

Alt-metrics total citations from all sources.
}
\keyword{internal}

