% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{package}
\name{alm-package}
\alias{alm}
\alias{alm-package}
\title{R client for the open source article level metrics ALM application.}
\description{
R client for the open source article level metrics ALM application.
}
\details{
An R interface to the open source article level metrics API
created by the Public Library of Science (PLOS). A number of publishers are
using the open source app created by PLOS, so you can drop in a different
base URL to the functions in this package to get to not only PLOS data, but
data for Crossref, and more as the open source PLOS software is used.

You need API keys for each of the endpoints you use.
}
\author{
Scott Chamberlain \email{myrmecocystus@gmail.com}
}
\keyword{package}

