% Generated by roxygen2 (4.0.1): do not edit by hand
\name{alm}
\alias{alm}
\title{Retrieve PLoS article-level metrics (ALM).}
\usage{
alm(doi = NULL, pmid = NULL, pmcid = NULL, mendeley = NULL,
  info = "totals", source = NULL, key = NULL, total_details = FALSE,
  sum_metrics = NULL, url = "http://alm.plos.org/api/v5/articles", ...)
}
\arguments{
\item{doi}{Digital object identifier for an article in PLoS Journals (character)}

\item{pmid}{PubMed object identifier (numeric)}

\item{pmcid}{PubMed Central object identifier (numeric)}

\item{mendeley}{Mendeley object identifier (character)}

\item{url}{API endpoint, defaults to http://alm.plos.org/api/v3/articles (character)}

\item{info}{One of totals, summary, or detail (default totals + sum_metrics data in a list).
Not specifying anything (the default) returns data.frame of totals across
data providers. (character)}

\item{source}{Name of source (or list of sources) to get ALM information for (character)}

\item{key}{your PLoS API key, either enter, or loads from .Rprofile (character)}

\item{...}{optional additional curl options (debugging tools mostly)}

\item{total_details}{If FALSE (the default) the standard totals data.frame is
returned; if TRUE, the totals data is in a wide format with more details
about the paper, including publication date, title, etc. If you set this
to TRUE, the output should no longer with with \code{\link{almplot}}.}

\item{sum_metrics}{Just like the output you get from setting info='totals', you can
get summary metrics by day (sum_metrics='day'), month (sum_metrics='month'),
or year (sum_metrics='year').}
}
\value{
PLoS altmetrics as data.frame's.
}
\description{
This is the main function to search the PLoS ALM (article-level metrics) API.
See details for more information.
}
\details{
You can only supply one of the parmeters doi, pmid, pmcid, and mendeley.

		Query for as many articles at a time as you like. Though queries are broken
		up in to smaller bits of 50 identifiers at a time.

		If you supply days, months and/or year parameters, days takes precedence
		over months and year.
}
\examples{
\dontrun{
# The default call with either doi, pmid, pmcid, or mendeley without specifying
# an argument for info
alm(doi="10.1371/journal.pone.0029797")

# Details for a single DOI
out <- alm(doi='10.1371/journal.pone.0029797', info='detail')
## totals
out[["totals"]]
## history
out[["history"]]

# A single PubMed ID (pmid)
alm(pmid=22590526)

# A single PubMed Central ID (pmcid)
alm(pmcid=212692, info='summary')

# A single Mendeley UUID (mendeley)
alm(mendeley="437b07d9-bc40-4c57-b60e-1f60fefe2300")
alm(mendeley=c("edc2e519-cc10-36fc-a68d-12e0116c6ac0", "62128d98-b63d-3f26-9bb4-0bda3913f01e"))

# Provide more than one DOI
dois <- c('10.1371/journal.pone.0001543','10.1371/journal.pone.0040117',
		'10.1371/journal.pone.0029797','10.1371/journal.pone.0039395')
out <- alm(doi=dois)
out[[1]] # get data for the first DOI

# Search for DOI's, then feed into alm
library('rplos')
dois <- searchplos(q='evolution', fl='id',
   fq=list('-article_type:correction','doc_type:full'), limit = 52)
out <- alm(doi=as.character(dois[,1]))
lapply(out, head)

# Provide more than one pmid
pmids <- c(19300479, 19390606, 19343216)
out <- alm(pmid=pmids)
out[[3]] # get data for the third pmid

# Getting data for a specific source
alm(doi='10.1371/journal.pone.0035869', source='mendeley')
alm(doi='10.1371/journal.pone.0035869', source=c('mendeley','twitter','counter'))
alm(doi='10.1371/journal.pone.0035869', source=c('mendeley','twitter','counter'), info='history')

# Get detailed totals output
alm(doi='10.1371/journal.pone.0035869', total_details=TRUE)

# Get summary metrics by day
alm(doi='10.1371/journal.pone.0036240', sum_metrics='day')

# Get summary metrics by month
alm(doi='10.1371/journal.pone.0036240', sum_metrics='month')

# Get summary metrics by year
alm(doi='10.1371/journal.pone.0036240', sum_metrics='year')

# Curl debugging
library('httr')
alm(doi="10.1371/journal.pone.0029797", config=verbose())
dois <- c('10.1371/journal.pone.0001543','10.1371/journal.pone.0040117',
 	'10.1371/journal.pone.0029797','10.1371/journal.pone.0039395')
alm(doi=dois, config=verbose())
}
}
\references{
See a tutorial/vignette for alm at
\url{http://ropensci.org/tutorials/alm_tutorial.html}
}
\seealso{
\code{\link{almplot}}
}

