% Generated by roxygen2 (4.0.1): do not edit by hand
\name{alm_pubmedcentid}
\alias{alm_pubmedcentid}
\title{Get PubMed Central article ID by inputting the doi for the article.}
\usage{
alm_pubmedcentid(doi, key = NULL, curl = getCurlHandle())
}
\arguments{
\item{doi}{digital object identifier for an article in PLoS Journals}

\item{key}{your PLoS API key, either enter, or loads from .Rprofile}

\item{curl}{If using in a loop, call getCurlHandle() first and pass
the returned value in here (avoids unnecessary footprint)}
}
\value{
The PubMed Central article ID.
}
\description{
Get PubMed Central article ID by inputting the doi for the article.
}
\examples{
\dontrun{
alm_pubmedcentid(doi = '10.1371/journal.pbio.0000012')
}
}
\references{
See a tutorial/vignette for alm at
\url{http://ropensci.org/tutorials/alm_tutorial.html}
}

