% Generated by roxygen2 (4.0.1): do not edit by hand
\name{alm_pubmedid}
\alias{alm_pubmedid}
\title{Get PubMed article ID by inputting the doi for the article.}
\usage{
alm_pubmedid(doi, key = NULL)
}
\arguments{
\item{doi}{digital object identifier for an article in PLoS Journals}

\item{key}{your PLoS API key, either enter, or loads from .Rprofile}
}
\value{
The PubMed article ID.
}
\description{
Get PubMed article ID by inputting the doi for the article.
}
\examples{
\dontrun{
alm_pubmedid(doi = '10.1371/journal.pbio.0000012')
}
}
\references{
See a tutorial/vignette for alm at
\url{http://ropensci.org/tutorials/alm_tutorial.html}
}

