% Generated by roxygen2 (4.0.2): do not edit by hand
\name{alm_requests}
\alias{alm_requests}
\title{Retrieve API requests data.}
\usage{
alm_requests(key = NULL, url = "http://alm.plos.org/api/v5/api_requests",
  ...)
}
\arguments{
\item{key}{API key}

\item{url}{URL for the api requests endpoint}

\item{...}{Curl args to \code{\link[httr]{GET}}}
}
\description{
Retrieve API requests data.
}
\examples{
\donttest{
out <- alm_requests()
out$meta
dat <- out$data
head( dat )
dat[ dat$db_duration > 100, ]
}
}

