% Generated by roxygen2 (4.0.1): do not edit by hand
\name{alm_source}
\alias{alm_source}
\title{Retrieve PLoS article-level metrics (ALM) by source.}
\usage{
alm_source(source = "crossref", info = "totals", key = NULL,
  total_details = FALSE, sum_metrics = NULL, limit = 50, page = 1,
  url = "http://alm.plos.org/api/v5/articles", ...)
}
\arguments{
\item{source}{(character) Name of source (or list of sources) to get ALM information for. Can be
one or more sources in a vector or list.}

\item{info}{One of totals, summary, or detail (default totals + sum_metrics data in a list).
Not specifying anything (the default) returns data.frame of totals across
  data providers. (character)}

\item{key}{your PLoS API key, either enter, or loads from .Rprofile (character)}

\item{total_details}{If FALSE (the default) the standard totals data.frame is
returned; if TRUE, the totals data is in a wide format with more details
about the paper, including publication date, title, etc. If you set this
to TRUE, the output should no longer with with \code{\link{almplot}}.}

\item{sum_metrics}{Just like the output you get from setting info='totals', you can
get summary metrics by day (sum_metrics='day'), month (sum_metrics='month'),
or year (sum_metrics='year').}

\item{url}{API endpoint, defaults to http://alm.plos.org/api/v3/articles (character)}

\item{...}{optional additional curl options (debugging tools mostly)}
}
\description{
See details for more information.
}
\examples{
\dontrun{
alm_source(source='mendeley')
alm_source(source='scopus', info='summary')
alm_source(source=c('mendeley','twitter'))
alm_source(source='mendeley', limit=2)
alm_source(source='mendeley', limit=200)

alm_source(source='mendeley', info='summary')
}
}
\references{
See a tutorial/vignette for alm at
\url{http://ropensci.org/tutorials/alm_tutorial.html}
}

