% Generated by roxygen2 (4.0.2): do not edit by hand
\name{alm_status}
\alias{alm_status}
\title{Check status of an ALM service.}
\usage{
alm_status(key = NULL, url = "http://alm.plos.org/api/v5/status", ...)
}
\arguments{
\item{key}{your PLoS API key, either enter, or loads from .Rprofile (character)}

\item{url}{API endpoint, defaults to http://alm.plos.org/api/v5/status (character)}

\item{...}{optional additional curl options (debugging tools mostly)}
}
\description{
Check status of an ALM service.
}
\examples{
\dontrun{
alm_status()
}
}

