% Generated by roxygen2 (4.0.1): do not edit by hand
\name{alm_title}
\alias{alm_title}
\title{Get title of article by inputting the doi for the article.}
\usage{
alm_title(doi = NULL, pmid = NULL, pmcid = NULL, mendeley_uuid = NULL,
  key = NULL, url = "http://alm.plos.org/api/v5/articles", ...)
}
\arguments{
\item{doi}{digital object identifier for an article in PLoS Journals}

\item{key}{your PLoS API key, either enter, or loads from .Rprofile}
}
\value{
Title of article, in xml format.
}
\description{
Get title of article by inputting the doi for the article.
}
\examples{
\dontrun{
alm_title(doi='10.1371/journal.pbio.0000012')
dois <- c('10.1371/journal.pone.0026871','10.1371/journal.pone.0048868',
  	'10.1371/journal.pone.0048705','10.1371/journal.pone.0048731')
alm_title(doi=dois)
}
}
\references{
See a tutorial/vignette for alm at
\url{http://ropensci.org/tutorials/alm_tutorial.html}
}

