% Generated by roxygen2 (4.0.1): do not edit by hand
\name{alm_totals}
\alias{alm_totals}
\title{Alt-metrics total citations from all sources.}
\usage{
alm_totals(doi, key = NULL)
}
\arguments{
\item{doi}{digital object identifier for an article in PLoS Journals}

\item{key}{your PLoS API key, either enter, or loads from .Rprofile}
}
\value{
data.frame of total no. views (counter + pmc), shares (facebook + twitter),
		bookmarks (mendeley + citeulike), and citations (crossref)
}
\description{
Alt-metrics total citations from all sources.
}
\examples{
\dontrun{
alm_totals(doi = '10.1371/journal.pbio.0000012')
}
}
\references{
See a tutorial/vignette for alm at
\url{http://ropensci.org/tutorials/alm_tutorial.html}
}

