% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alm_works.R
\name{alm_works}
\alias{alm_work}
\alias{alm_works}
\title{Works}
\usage{
alm_works(ids = NULL, q = NULL, type = NULL, source_id = NULL,
  publisher_id = NULL, contributor_id = NULL, relation_type_id = NULL,
  registration_agency = NULL, sort = NULL, page = NULL, per_page = NULL,
  api_url = "https://eventdata.datacite.org", ...)

alm_work(id, api_url = "https://eventdata.datacite.org", ...)
}
\arguments{
\item{ids}{(character) Work IDs. optional}

\item{q}{(character) Query for ids. optional}

\item{type}{(character) Work ID type (one of doi, pmid, pmcid, arxiv, wos, 
scp, ark, or url). optional}

\item{source_id}{(character) Source ID. optional}

\item{publisher_id}{(character) Publisher ID. optional}

\item{contributor_id}{(character) Contributor ID. optional}

\item{relation_type_id}{(character) Relation_type ID. optional}

\item{registration_agency}{(character) Registration agency. optional}

\item{sort}{(character) Sort by source result count descending, or by publication 
date descending if left empty. optional}

\item{page}{(numeric) Page number. optional}

\item{per_page}{(numeric) Results per page (0-1000), defaults to 1000. optional}

\item{api_url}{(character) API endpoint, defaults to http://alm.plos.org/api/v5/articles}

\item{...}{Curl options passed on to \code{\link[httr]{GET}}}

\item{id}{(character) A single work ID. required}
}
\description{
Works
}
\details{
Returns list of works either by ID, or all

If no ids are provided in the query, all works are returned, 1000 per page and 
sorted by publication date (default), or source result count. Search is not 
supported by the API.
}
\examples{
\dontrun{
alm_work(id = 'http://doi.org/10.6068/DP1541A4866AD14')

alm_works(ids = 'http://doi.org/10.6068/DP1541A4866AD14')
ids <- c('http://doi.org/10.6068/DP1541A53768228','http://doi.org/10.6068/DP1541A4866AD14')
alm_works(ids = ids)
}
}

