% Generated by roxygen2 (4.0.1): do not edit by hand
\name{almtitle}
\alias{almtitle}
\title{Get title of article by inputting the doi for the article.}
\usage{
almtitle(doi, key = NULL)
}
\arguments{
\item{doi}{digital object identifier for an article in PLoS Journals}

\item{key}{your PLoS API key, either enter, or loads from .Rprofile}
}
\value{
Title of article, in xml format.
}
\description{
Get title of article by inputting the doi for the article.
}
\examples{
\dontrun{
almtitle(doi='10.1371/journal.pbio.0000012')
}
}
\references{
See a tutorial/vignette for alm at
\url{http://ropensci.org/tutorials/alm_tutorial.html}
}

