% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_requests.R
\name{api_requests}
\alias{api_requests}
\title{API requests}
\usage{
api_requests(key, api_url = "http://alm.plos.org", ...)
}
\arguments{
\item{key}{(character) An API key. Required.}

\item{api_url}{(character) API endpoint, defaults to http://alm.plos.org/api/v5/articles}

\item{...}{Curl options (debugging tools mostly) passed on to \code{\link[httr]{GET}}}
}
\description{
API requests
}
\examples{
\dontrun{
api_requests(getOption("almv4_pwd"))
}
}

