% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contributions.R
\name{contributions}
\alias{contributions}
\title{Contributions}
\usage{
contributions(contributor_role_id = NULL, source_id = NULL, recent = NULL,
  page = NULL, per_page = NULL,
  api_url = "https://eventdata.datacite.org", ...)
}
\arguments{
\item{contributor_role_id}{(character) Contributor_role ID. optional}

\item{source_id}{(character) Source name. optional}

\item{recent}{(integer) Limit to relations created last x days. optional}

\item{page}{(numeric) Page number. optional}

\item{per_page}{(numeric) Results per page, defaults to 1000. optional}

\item{api_url}{(character) API endpoint, defaults to http://alm.plos.org/api/v5/articles}

\item{...}{Curl options passed on to \code{\link[httr]{GET}}}
}
\description{
Contributions
}
\examples{
\dontrun{
res <- contributions()
res$meta
head(res$contributions)
}
}

