% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contributor_roles.R
\name{contributor_roles}
\alias{contributor_roles}
\title{Contributor roles}
\usage{
contributor_roles(api_url = "https://eventdata.datacite.org", ...)
}
\arguments{
\item{api_url}{(character) API endpoint, defaults to http://alm.plos.org/api/v5/articles}

\item{...}{Curl options passed on to \code{\link[httr]{GET}}}

\item{id}{(character) contributor ID. optional}
}
\description{
Contributor roles
}
\examples{
\dontrun{
res <- contributor_roles()
res$meta
head(res$contributor_roles)

contributor_role('conceptualization')
}
}

