% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contributors.R
\name{contributors}
\alias{contributor}
\alias{contributors}
\title{Contributors}
\usage{
contributors(page = NULL, per_page = NULL,
  api_url = "https://eventdata.datacite.org", ...)

contributor(id, api_url = "https://eventdata.datacite.org", ...)
}
\arguments{
\item{page}{(numeric) Page number. optional}

\item{per_page}{(numeric) Results per page, defaults to 1000. optional}

\item{api_url}{(character) API endpoint, defaults to http://alm.plos.org/api/v5/articles}

\item{...}{Curl options passed on to \code{\link[httr]{GET}}}

\item{id}{(character) contributor ID. optional}
}
\description{
Contributors
}
\examples{
\dontrun{
res <- contributors()
res$meta
head(res$contributors)

contributors(per_page = 2)

contributor('https://github.com/golo-lang')
}
}

