% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_exports.R
\name{data_exports}
\alias{data_exports}
\title{Data exports}
\usage{
data_exports(key, api_url = "http://alm.plos.org", ...)
}
\arguments{
\item{key}{(character) An API key. Required.}

\item{api_url}{(character) API endpoint, defaults to http://alm.plos.org/api/v5/articles}

\item{...}{Curl options (debugging tools mostly) passed on to \code{\link[httr]{GET}}}
}
\description{
Data exports
}
\examples{
\dontrun{
data_exports(getOption("almv4_pwd"))
}
}

