% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deposits.R
\name{deposits}
\alias{deposit}
\alias{deposits}
\title{Deposits}
\usage{
deposits(message_type = NULL, state = NULL, source_token = NULL,
  page = NULL, per_page = NULL,
  api_url = "https://eventdata.datacite.org", ...)

deposit(id, api_url = "https://eventdata.datacite.org", ...)
}
\arguments{
\item{message_type}{(character) Filter by message_type. optional}

\item{state}{(character) "Filter by DOI prefix" or maybe "Filter by state" optional}

\item{source_token}{(character) Filter by source_token. optional}

\item{page}{(numeric) Page number. optional}

\item{per_page}{(numeric) Results per page, defaults to 1000. optional}

\item{api_url}{(character) API endpoint, defaults to http://alm.plos.org/api/v5/articles}

\item{...}{Curl options passed on to \code{\link[httr]{GET}}}

\item{id}{(character) A deposit id, required.}
}
\description{
Deposits
}
\examples{
\dontrun{
res <- deposits()
res$meta
head(res$deposits)

deposits(per_page = 2)

# message type
deposits(message_type = 'relation')

# source token
deposits(source_token = "5f9f3976-415f-4fb7-b5df-b7e0329d1258")

# source token
deposit(id = "67495ea3-f9d2-40cf-8336-10c87bb3ae83")
}
}

