% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docs.R
\name{docs}
\alias{docs}
\title{Documents}
\usage{
docs(id = NULL, api_url = "http://alm.plos.org", ...)
}
\arguments{
\item{id}{(character) A group id. Optional.}

\item{api_url}{(character) API endpoint, defaults to http://alm.plos.org/api/v5/articles}

\item{...}{Curl options (debugging tools mostly) passed on to \code{\link[httr]{GET}}}
}
\description{
Documents
}
\examples{
\dontrun{
docs()
docs('wos')
docs('crossref')
}
}

