% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groups.R
\name{groups}
\alias{group}
\alias{groups}
\title{Groups}
\usage{
groups(api_url = "https://eventdata.datacite.org", ...)

group(id, api_url = "https://eventdata.datacite.org", ...)
}
\arguments{
\item{api_url}{(character) API endpoint, defaults to http://alm.plos.org/api/v5/articles}

\item{...}{Curl options passed on to \code{\link[httr]{GET}}}

\item{id}{(character) A group id. Optional.}
}
\description{
Groups
}
\examples{
\dontrun{
groups()
group('contributions')
group('works')
}
}

