% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/notifications.R
\name{notifications}
\alias{notifications}
\title{Notifications}
\usage{
notifications(key, api_url = "https://eventdata.datacite.org", ...)
}
\arguments{
\item{key}{(character) An API key. Required.}

\item{api_url}{(character) API endpoint, defaults to http://alm.plos.org/api/v5/articles}

\item{...}{Curl options passed on to \code{\link[httr]{GET}}}
}
\description{
Notifications
}

