% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/publishers.R
\name{publishers}
\alias{publisher}
\alias{publishers}
\title{Publishers}
\usage{
publishers(api_url = "https://eventdata.datacite.org", ...)

publisher(id, api_url = "https://eventdata.datacite.org", ...)
}
\arguments{
\item{api_url}{(character) API endpoint, defaults to http://alm.plos.org/api/v5/articles}

\item{...}{Curl options passed on to \code{\link[httr]{GET}}}

\item{id}{(character) A publisher id. required}
}
\description{
Publishers
}
\examples{
\dontrun{
publishers()
publisher(340)
}
}

