% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recommendations.R
\name{recommendations}
\alias{recommendations}
\title{Recommendations}
\usage{
recommendations(doi, api_url = "https://eventdata.datacite.org", ...)
}
\arguments{
\item{doi}{(character) A doi. Required.}

\item{api_url}{(character) API endpoint, defaults to http://alm.plos.org/api/v5/articles}

\item{...}{Curl options passed on to \code{\link[httr]{GET}}}
}
\description{
Recommendations
}
\examples{
\dontrun{
res <- recommendations("http://doi.org/10.1038/NGEO2696")
res$meta
res$recommendations
}
}

