% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/references.R
\name{references}
\alias{references}
\title{References}
\usage{
references(doi, api_url = "http://alm.plos.org", ...)
}
\arguments{
\item{doi}{(character) A DOI. Required.}

\item{api_url}{(character) API endpoint, defaults to http://alm.plos.org/api/v5/articles}

\item{...}{Curl options (debugging tools mostly) passed on to \code{\link[httr]{GET}}}
}
\description{
References
}
\examples{
\dontrun{
res <- references("10.1371/journal.pone.0017342")
res$meta
res$references
}
}

