% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/relation_types.R
\name{relation_types}
\alias{relation_types}
\title{Relation types}
\usage{
relation_types(x = NULL, api_url = "http://alm.plos.org", ...)
}
\arguments{
\item{x}{(character) A relation type name. Optional.}

\item{api_url}{(character) API endpoint, defaults to http://alm.plos.org/api/v5/articles}

\item{...}{Curl options (debugging tools mostly) passed on to \code{\link[httr]{GET}}}
}
\description{
Relation types
}
\examples{
\dontrun{
res <- relation_types()
res$meta
res$relation_types

relation_types("bookmarks")
relation_types("discusses")
}
}

