% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relations.R
\name{relations}
\alias{relations}
\title{Relations}
\usage{
relations(work_id = NULL, work_ids = NULL, q = NULL,
  relation_type_id = NULL, source_id = NULL, recent = NULL, page = NULL,
  per_page = NULL, api_url = "https://eventdata.datacite.org", ...)
}
\arguments{
\item{work_id}{(character) Work ID. optional}

\item{work_ids}{(character) Work IDs. optional}

\item{q}{(character) Query for ids. optional}

\item{relation_type_id}{(character) Relation_type ID. optional}

\item{source_id}{(character) Source name. optional}

\item{recent}{(integer) Limit to relations created last x days. optional}

\item{page}{(numeric) Page number. optional}

\item{per_page}{(numeric) Results per page, defaults to 1000. optional}

\item{api_url}{(character) API endpoint, defaults to http://alm.plos.org/api/v5/articles}

\item{...}{Curl options passed on to \code{\link[httr]{GET}}}
}
\description{
Relations
}
\examples{
\dontrun{
res <- relations()
res$meta
head(res$relations)
}
}

