% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results.R
\name{results}
\alias{results}
\title{Results}
\usage{
results(work_id = NULL, work_ids = NULL, source_id = NULL, sort = NULL,
  page = NULL, per_page = NULL,
  api_url = "https://eventdata.datacite.org", ...)
}
\arguments{
\item{work_id}{(character) Work ID. optional}

\item{work_ids}{(character) Work IDs. optional}

\item{source_id}{(character) Source name. optional}

\item{sort}{(character) Sort by result descending, or by update date descending 
if left empty. optional}

\item{page}{(numeric) Page number. optional}

\item{per_page}{(numeric) Results per page, defaults to 1000. optional}

\item{api_url}{(character) API endpoint, defaults to http://alm.plos.org/api/v5/articles}

\item{...}{Curl options passed on to \code{\link[httr]{GET}}}
}
\description{
Results
}
\examples{
\dontrun{
res <- results()
res$meta
head(res$results)
}
}

