% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sources.R
\name{sources}
\alias{source_}
\alias{sources}
\alias{sources_months}
\title{Sources}
\usage{
sources(api_url = "https://eventdata.datacite.org", ...)

source_(id, api_url = "https://eventdata.datacite.org", ...)

sources_months(x, api_url = "https://eventdata.datacite.org", ...)
}
\arguments{
\item{api_url}{(character) API endpoint, defaults to http://alm.plos.org/api/v5/articles}

\item{...}{Curl options passed on to \code{\link[httr]{GET}}}

\item{id}{(character) A source id. required}
}
\description{
Sources
}
\examples{
\dontrun{
res <- sources()
res$meta
res$sources

source("crossref_datacite")
source("datacite_orcid")

# months data
sources_months("crossref_datacite")
}
}

