% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/work_types.R
\name{work_types}
\alias{work_type}
\alias{work_types}
\title{Work types}
\usage{
work_types(api_url = "https://eventdata.datacite.org", ...)

work_type(id, api_url = "https://eventdata.datacite.org", ...)
}
\arguments{
\item{api_url}{(character) API endpoint, defaults to http://alm.plos.org/api/v5/articles}

\item{...}{Curl options passed on to \code{\link[httr]{GET}}}

\item{id}{A work type id, required}
}
\description{
Work types
}
\examples{
\dontrun{
res <- work_types()
res$meta
res$work_types

work_type("speech")
}
}

