% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/alm-package.R, R/defunct.R
\name{alm-defunct}
\alias{alm-defunct}
\alias{alm_pubmedcentid}
\alias{alm_pubmedid}
\alias{alm_totals}
\alias{almdateupdated}
\alias{almupdated}
\title{Defunct functions in alm}
\usage{
alm_pubmedid(...)

alm_pubmedcentid(...)

almupdated(...)

almdateupdated(...)

alm_totals(...)
}
\arguments{
\item{...}{Args}
}
\description{
\itemize{
 \item \code{\link{alm_pubmedid}}: Function removed, you can get this info using alm_ids
 \item \code{\link{alm_pubmedcentid}}: Function removed, you can get this info using alm_ids
 \item \code{\link{almupdated}}: Function removed, you can get this info using alm_ids
 \item \code{\link{almdateupdated}}: Function removed, you can get this info using alm_ids
 \item \code{\link{alm_totals}}: Function removed, you can get this info using alm_ids
}

Get PubMed article ID by inputting the doi for the article.

Get PubMed Central article ID by inputting the doi for the article.

Date when alt-metrics for the article (by DOI) data was last updated.

Get the date when article was last updated.

Alt-metrics total citations from all sources.
}

