% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/alm_status.R
\name{alm_status}
\alias{alm_status}
\title{Check status of an ALM service.}
\usage{
alm_status(key = NULL, url = "http://alm.plos.org/api/v5/status", ...)
}
\arguments{
\item{key}{(character) Your API key, either enter, or loads from .Rprofile. Only required for
PKP source, not the others.}

\item{url}{API endpoint, defaults to http://alm.plos.org/api/v5/status (character)}

\item{...}{optional additional curl options (debugging tools mostly)}
}
\description{
Check status of an ALM service.
}
\examples{
\dontrun{
alm_status()
}
}

