% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alm_works.R
\name{alm_works}
\alias{alm_works}
\title{ALM works}
\usage{
alm_works(doi = NULL, pmid = NULL, pmcid = NULL, wos = NULL,
  scp = NULL, url = NULL, arxiv = NULL, ark = NULL, source_id = NULL,
  publisher_id = NULL, api_url = "http://alm.plos.org", ...)
}
\arguments{
\item{doi}{xxx}

\item{pmid}{xxx}

\item{pmcid}{xxx}

\item{wos}{xxx}

\item{scp}{xxx}

\item{url}{xxx}

\item{arxiv}{xxx}

\item{ark}{xxx}

\item{source_id}{xxx}

\item{publisher_id}{xxx}

\item{api_url}{(character) API endpoint, defaults to http://alm.plos.org/api/v5/articles}

\item{...}{Curl options (debugging tools mostly) passed on to \code{\link[httr]{GET}}}
}
\description{
ALM works
}
\examples{
\dontrun{
alm_works(doi = '10.1371/journal.ppat.1005123')
}
}

