% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contributor_roles.R, R/contributors.R, R/docs.R
\name{contributor_role}
\alias{contributor}
\alias{contributor_role}
\alias{doc}
\alias{docs}
\title{Documents}
\usage{
contributor_role(id, api_url = "https://eventdata.datacite.org", ...)

contributor(id, api_url = "https://eventdata.datacite.org", ...)

docs(id = NULL, api_url = "https://eventdata.datacite.org", ...)

doc(id = NULL, api_url = "https://eventdata.datacite.org", ...)
}
\arguments{
\item{id}{(character) A group id. Optional.}

\item{api_url}{(character) API endpoint, defaults to http://alm.plos.org/api/v5/articles}

\item{...}{Curl options passed on to \code{\link[httr]{GET}}}
}
\description{
Documents
}
\examples{
\dontrun{
docs()
doc('wos')
doc('crossref')
}
}

