% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groups.R
\name{groups}
\alias{groups}
\title{Groups}
\usage{
groups(id = NULL, api_url = "http://alm.plos.org", ...)
}
\arguments{
\item{id}{(character) A group id. Optional.}

\item{api_url}{(character) API endpoint, defaults to http://alm.plos.org/api/v5/articles}

\item{...}{Curl options (debugging tools mostly) passed on to \code{\link[httr]{GET}}}
}
\description{
Groups
}
\examples{
\dontrun{
groups()
groups('cited')
groups('viewed')
}
}

