% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/publishers.R
\name{publishers}
\alias{publishers}
\title{Publishers}
\usage{
publishers(id = NULL, api_url = "http://alm.plos.org", ...)
}
\arguments{
\item{id}{(character) A publisher id. Optional.}

\item{api_url}{(character) API endpoint, defaults to http://alm.plos.org/api/v5/articles}

\item{...}{Curl options (debugging tools mostly) passed on to \code{\link[httr]{GET}}}
}
\description{
Publishers
}
\examples{
\dontrun{
publishers()
publishers(340)
}
}

