% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/recommendations.R
\name{recommendations}
\alias{recommendations}
\title{Recommendations}
\usage{
recommendations(doi, api_url = "http://alm.plos.org", ...)
}
\arguments{
\item{doi}{(character) A doi. Required.}

\item{api_url}{(character) API endpoint, defaults to http://alm.plos.org/api/v5/articles}

\item{...}{Curl options (debugging tools mostly) passed on to \code{\link[httr]{GET}}}
}
\description{
Recommendations
}
\examples{
\dontrun{
res <- recommendations("10.1371/journal.pmed.1001361")
res$meta
res$recommendations
}
}

