% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sources.R
\name{sources}
\alias{sources}
\alias{sources_months}
\title{Sources}
\usage{
sources(x = NULL, api_url = "http://alm.plos.org", ...)

sources_months(x, api_url = "http://alm.plos.org", ...)
}
\arguments{
\item{x}{(character) A source name. Optional for \code{sources}; required
for \code{sources_months}.}

\item{api_url}{(character) API endpoint, defaults to http://alm.plos.org/api/v5/articles}

\item{...}{Curl options (debugging tools mostly) passed on to \code{\link[httr]{GET}}}
}
\description{
Sources
}
\examples{
\dontrun{
res <- sources()
res$meta
res$sources

sources("citeulike")
sources("crossref")

# months data
sources_months("crossref")
}
}

