% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/status.R
\name{status}
\alias{status}
\title{Check the status of a Lagotto API}
\usage{
status(api_url = "https://eventdata.datacite.org", ...)
}
\arguments{
\item{api_url}{(character) API endpoint, defaults to http://alm.plos.org/api/v5/articles}

\item{...}{Curl options passed on to \code{\link[httr]{GET}}}
}
\description{
Check the status of a Lagotto API
}
\examples{
\dontrun{
res <- status()
res$meta
res$status
head(res$status$agents)
}
}

